requirements.txt
flask
numpy
pandas 
gunicorn
 pickle
 streamlit 
 requests
 pillow

 colorama
 rich